da sistemare pagina admin

+++ WaterManager/src/main/java/pissir/watermanager/controller/ControllerRichiestaIdrica.java
@@ -2,12 +2,11 @@ package pissir.watermanager.controller;
 
 import com.google.gson.Gson;
 import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.security.access.prepost.PreAuthorize;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
 import pissir.watermanager.dao.DAO;
 import pissir.watermanager.model.item.RichiestaIdrica;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import pissir.watermanager.model.user.UserRole;
 import pissir.watermanager.security.services.TokenService;
 
 import java.util.HashSet;
@@ -26,29 +25,53 @@ public class ControllerRichiestaIdrica {
 	
 	
 	@PostMapping(value = "/add")
-	public ResponseEntity<Integer> addRichiesta (@RequestBody String param) {
+	public String addRichiesta(@RequestBody String param, HttpServletRequest request) {
 		Gson gson = new Gson();
-		RichiestaIdrica richiesta = gson.fromJson(param, RichiestaIdrica.class);
+		String jwt = extractTokenFromRequest(request);
 		
-		return ResponseEntity.ok(this.daoRichieste.addRichiesta(richiesta));
+		if (this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREIDRICO) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREAZIENDA) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.SYSTEMADMIN)) {
+			RichiestaIdrica richiesta = gson.fromJson(param, RichiestaIdrica.class);
+			
+			return gson.toJson(this.daoRichieste.addRichiesta(richiesta));
+		} else {
+			return gson.toJson("Accesso negato");
+		}
 	}
 	
 	
 	@GetMapping(value = "/get/{id}")
-	public String getRichiestaId (@PathVariable int id) {
+	public String getRichiestaId(@PathVariable int id, HttpServletRequest request) {
 		Gson gson = new Gson();
-		RichiestaIdrica richiestaIdrica = this.daoRichieste.getRichiestaId(id);
+		String jwt = extractTokenFromRequest(request);
 		
-		return gson.toJson(richiestaIdrica);
+		if (this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREIDRICO) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREAZIENDA) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.SYSTEMADMIN)) {
+			RichiestaIdrica richiestaIdrica = this.daoRichieste.getRichiestaId(id);
+			
+			return gson.toJson(richiestaIdrica);
+		} else {
+			return gson.toJson("Accesso negato");
+		}
 	}
 	
 	
 	@GetMapping(value = "/get/allBacino/{id}")
-	public String getRichiesteBacino (@PathVariable int id) {
+	public String getRichiesteBacino(@PathVariable int id, HttpServletRequest request) {
 		Gson gson = new Gson();
-		HashSet<RichiestaIdrica> richieste = this.daoRichieste.getRichiesteBacino(id);
+		String jwt = extractTokenFromRequest(request);
 		
-		return gson.toJson(richieste);
+		if (this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREIDRICO) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREAZIENDA) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.SYSTEMADMIN)) {
+			HashSet<RichiestaIdrica> richieste = this.daoRichieste.getRichiesteBacino(id);
+			
+			return gson.toJson(richieste);
+		} else {
+			return gson.toJson("Accesso negato");
+		}
 	}
 
 	/*
@@ -64,11 +87,19 @@ public class ControllerRichiestaIdrica {
 	
 	
 	@GetMapping(value = "/get/allColtivazione/{id}")
-	public String getRichiesteColtivazione (@PathVariable int id) {
+	public String getRichiesteColtivazione(@PathVariable int id, HttpServletRequest request) {
 		Gson gson = new Gson();
-		HashSet<RichiestaIdrica> richieste = this.daoRichieste.getRichiesteColtivazione(id);
+		String jwt = extractTokenFromRequest(request);
 		
-		return gson.toJson(richieste);
+		if (this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREIDRICO) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREAZIENDA) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.SYSTEMADMIN)) {
+			HashSet<RichiestaIdrica> richieste = this.daoRichieste.getRichiesteColtivazione(id);
+			
+			return gson.toJson(richieste);
+		} else {
+			return gson.toJson("Accesso negato");
+		}
 	}
 
 	/*
@@ -82,8 +113,18 @@ public class ControllerRichiestaIdrica {
 	
 	
 	@DeleteMapping("/delete/{id}")
-	public void deleteRichiesta (@PathVariable int id) {
-		this.daoRichieste.deleteRichiesta(id);
+	public String deleteRichiesta(@PathVariable int id, HttpServletRequest request) {
+		Gson gson = new Gson();
+		String jwt = extractTokenFromRequest(request);
+		
+		if (this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREIDRICO) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.GESTOREAZIENDA) ||
+				this.tokenService.validateTokenAndRole(jwt, UserRole.SYSTEMADMIN)) {
+			this.daoRichieste.deleteRichiesta(id);
+			return "OK";
+		} else {
+			return gson.toJson("Accesso negato");
+		}
 	}
 	
 	
@@ -94,4 +135,5 @@ public class ControllerRichiestaIdrica {
 		}
 		return null;
 	}
+	
 }
